requirements.txt
asgiref==3.3.1
boto3==1.16.35
botocore==1.19.35
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-js-asset==1.2.2
django-mailer==2.1
django-mptt==0.11.0
django-storages==1.10.1
djangorestframework==3.12.2
idna==2.10
jmespath==0.10.0
lockfile==0.12.2
Pillow==8.0.1
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
wincertstore==0.2 
django-redis==4.12.1